back/requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
asyncpg==0.30.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
Django==5.1.3
django-extensions==3.2.3
django-phonenumber-field==8.0.0
django-unfold==0.49.1
djangorestframework==3.15.2
easydict==1.13
eval_type_backport==0.2.0
executing==2.1.0
flatbuffers==24.3.25
fonttools==4.55.0
h11==0.14.0
humanfriendly==10.0
idna==3.10
imageio==2.36.1
insightface==0.7.3
ipython==8.30.0
jedi==0.19.2
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.9.3
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
onnx==1.17.0
onnxruntime==1.20.1
opencv-python-headless==4.10.0.84
packaging==24.2
parso==0.8.4
pgvector==0.3.6
phonenumbers==8.13.52
pillow==11.0.0
prettytable==3.12.0
prompt_toolkit==3.0.48
protobuf==5.28.2
psycopg==3.2.6
psycopg-binary==3.2.6
pure_eval==0.2.3
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rdflib==7.1.3
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.6.0
simsimd==6.2.1
six==1.16.0
SPARQLWrapper==2.0.0
sqlparse==0.5.2
stack-data==0.6.3
stringzilla==3.10.11
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2024.9.20
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.9.0
django-filter~=25.1
python-decouple~=3.8
gunicorn==21.2.0
django-cors-headers==4.7.0